+++ config/wdio.conf.js
@@ -115,7 +115,7 @@ module.exports.configure = (options) => {
 				// maxInstances can get overwritten per capability. So if you have an in-house Selenium
 				// grid with only 5 firefox instances available you can make sure that not more than
 				// 5 instances get started at a time.
-				maxInstances,
+				'wdio:maxInstances': maxInstances,
 				//
 				browserName: 'chrome',
 				/* WebdriverIO v8.14 and above downloads and uses the latest Chrome version when running tests.