+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Change Log
 
+## [unreleased]
+
+* Fixed ui-tests errorShot saving failure by adding `await` to `browser.saveScreenshot()`.
+* Fixed ss-tests number of running instances on tv board by fixing maxInstances config property name.
+
 ## [2.0.0-alpha.2] (March 7, 2025)
 
 * Updated `webdriverio` and related dependencies to version 9.